[REFACTO] separate buttons & select menus handlers
