+++ src/helpers/index.js
@@ -47,6 +47,7 @@ import {
   getPollFromTitle,
   addPollChoices,
   addPollVoter,
+  isPollEmptyVotes,
   isThisChoicePollVoter,
   getPollVoteIndexes,
   getThisChoicePollIndex,
@@ -136,6 +137,7 @@ export {
   getPollFromTitle,
   addPollChoices,
   addPollVoter,
+  isPollEmptyVotes,
   isThisChoicePollVoter,
   getPollVoteIndexes,
   getThisChoicePollIndex,
